+++ changelog.md
@@ -4,6 +4,7 @@ TBD
 Features
 --------
 * Options to limit size of LLM prompts; cache LLM prompt data.
+* Suggest only tables/views that contain the given columns when provided in a SELECT/DELETE query.
 
 
 Bug Fixes